+++ apps/frontend/src/components/editor/EditorTitle.tsx
@@ -1,13 +1,10 @@
-import { useState } from "react";
 import Picker from "@emoji-mart/react";
 import { JSONContent } from "novel";
 
 import Emoji from "@/components/commons/emoji";
 
-import useYDocStore from "@/store/useYDocStore";
-import { useYText } from "@/hooks/useYText";
-import { useOptimisticUpdatePage } from "@/hooks/usePages";
 import { cn } from "@/lib/utils";
+import { useEditorTitle } from "@/hooks/useEditorTitle";
 
 interface EditorTitleProps {
   currentPage: number;
@@ -27,55 +24,16 @@ export default function EditorTitle({
   currentPage,
   pageContent,
 }: EditorTitleProps) {
-  const { ydoc } = useYDocStore();
-  const [title, setYTitle] = useYText(ydoc, "title", currentPage);
-  const [emoji, setYEmoji] = useYText(ydoc, "emoji", currentPage);
-  const [isEmojiPickerOpen, setIsEmojiPickerOpen] = useState(false);
-
-  const optimisticUpdatePageMutation = useOptimisticUpdatePage({
-    id: currentPage ?? 0,
-  });
-
-  // title 변경 -> invalidate page -> 새로운 pageContent + YText 타이틀로 update 요청
-  const handleTitleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setYTitle(e.target.value);
-
-    optimisticUpdatePageMutation.mutate({
-      pageData: { title: e.target.value, content: pageContent, emoji },
-    });
-  };
-
-  const handleEmojiClick = ({ native }: Emoji) => {
-    setYEmoji(native);
-
-    optimisticUpdatePageMutation.mutate({
-      pageData: { title, content: pageContent, emoji: native },
-    });
-
-    setIsEmojiPickerOpen(false);
-  };
-
-  const handleTitleEmojiClick = () => {
-    setIsEmojiPickerOpen(!isEmojiPickerOpen);
-  };
-
-  const handleEmojiOutsideClick = () => {
-    if (!isEmojiPickerOpen) return;
-
-    setIsEmojiPickerOpen(false);
-  };
-
-  const handleRemoveIconClick = (e: React.MouseEvent) => {
-    e.stopPropagation();
-
-    setYEmoji("");
-
-    optimisticUpdatePageMutation.mutate({
-      pageData: { title, content: pageContent, emoji: "" },
-    });
-
-    setIsEmojiPickerOpen(false);
-  };
+  const {
+    emoji,
+    title,
+    isEmojiPickerOpen,
+    handleTitleEmojiClick,
+    handleRemoveIconClick,
+    handleEmojiClick,
+    handleEmojiOutsideClick,
+    handleTitleChange,
+  } = useEditorTitle(currentPage, pageContent);
 
   return (
     <div className="flex flex-col gap-3">
